requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-groq
langchain-core
streamlit
requests

